move etherscanApi into a separate file in common
